Added Autonomous for Cargoship
Cargoship autonomous can be used from HAB lvl 1 or 2

+++ src/main/java/frc/robot/commands/DrivePosition.java
@@ -21,7 +21,7 @@ public class DrivePosition extends Command {
    * set the coding units are on the wheel
    */
   private static final double UNITS_PER_INCH=4096/CIRCUMFRANCE;
-  public DrivePosition(int distance) {
+  public DrivePosition(double distance) {
     // Use requires() here to declare subsystem dependencies
     // eg. requires(chassis);
